requirements.txt
Django==1.7.6
django-tastypie # REST api
django-countries # countries list

dj-database-url==0.3.0
dj-static==0.0.6
geopy==1.1.2
psycopg2==2.5.3
python-social-auth==0.2.1
south==1.0.0

#Pinax-project-account section
pinax-theme-bootstrap==5.5.1
django-user-accounts==1.0c9
metron==1.3.1
eventlog==0.8.0
django-jsonfield==0.9.12
django-simple-history>=1.5.4
django-bootstrap3>=5.1.0
django-sekizai
django-widget-tweaks
